Minor simplifications

+++ starlark/library.go
@@ -429,7 +429,7 @@ func chr(thread *Thread, _ *Builtin, args Tuple, kwargs []Tuple) (Value, error) 
 		return nil, fmt.Errorf("chr: Unicode code point U+%X out of range (>0x10FFFF)", i)
 	}
 	ret := String(string(rune(i)))
-	if err := thread.AddAllocs(int64(EstimateSize(ret))); err != nil {
+	if err := thread.AddAllocs(EstimateSize(ret)); err != nil {
 		return nil, err
 	}
 	return ret, nil
